\name{apply_pipeline}
\alias{apply_pipeline}
\title{Apply wrangling pipeline operation}
\usage{
apply_pipeline(data, operation = "None")
}
\arguments{
\item{data}{Input data frame.}
\item{operation}{Operation label from UI.}
}
\value{Data frame.}
\description{Applies selected data wrangling operation.}
